Revert "Merge branch 'back' into 'main'"
This reverts merge request !94

+++ backend/focus/src/main/java/com/bb/focus/db/entity/statistics/MajorPerApplicant.java
@@ -34,6 +34,6 @@ public class MajorPerApplicant {
 
   @NotNull
   @ColumnDefault("0")
-  private int applicantCount;
+  private Long applicantCount;
 
 }
